requirements.txt
altgraph==0.17.2
asgiref==3.5.0
auto-py-to-exe==2.18.2
backports.zoneinfo==0.2.1
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==5.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
decorator==4.4.2
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.0.3
django-allauth==0.50.0
django-debug-toolbar==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.3
drf-yasg2==1.19.4
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
greenlet==1.1.2
httplib2==0.21.0
idna==3.3
imageio==2.31.6
imageio-ffmpeg==0.4.9
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
moviepy==1.0.3
mysqlclient==2.1.0
numpy==1.24.4
oauthlib==3.2.0
openapi-codec==1.3.2
packaging==21.3
Pillow==9.1.0
proglog==0.1.10
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
PyJWT==2.3.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tk==0.1.0
tqdm==4.66.1
uritemplate==4.1.1
urllib3==1.26.9
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
